+++ src/test/java/org/openphc/cce/collector/service/EventIngestionServiceTest.java
@@ -17,7 +17,7 @@ import org.openphc.cce.collector.api.exception.CloudEventValidationException;
 import org.openphc.cce.collector.api.exception.DuplicateEventException;
 import org.openphc.cce.collector.api.exception.PayloadValidationException;
 import org.openphc.cce.collector.api.exception.KafkaPublishException;
-import org.openphc.cce.collector.config.KafkaTopicProperties;
+import org.openphc.cce.common.kafka.KafkaTopicProperties;
 import org.openphc.cce.collector.domain.model.InboundEventLog;
 import org.openphc.cce.collector.domain.model.enums.InboundStatus;
 import org.openphc.cce.collector.domain.model.enums.RejectionReason;